+++ intents/__init__.py
@@ -12,10 +12,10 @@ class SessionEntity:
         self._session = session
         return self
 
+from intents.language_codes import LanguageCode
+
 from intents.model.intent import Intent
 from intents.model.entity import Entity, Sys
 from intents.model.agent import Agent
 from intents.model.relations import follow
-from intents.model.fulfillment import FulfillmentRequest, FulfillmentResponse
-
-from intents.language import LanguageCode
+# from intents.model.fulfillment import FulfillmentContext, FulfillmentResult